Add files via upload
The project includes the creation of a different version (squares, rectangles, and circles) for Windows and Linux. It includes many basic capabilities for drawing shapes and providing descriptions, with the ability to obtain the properties of these electronic products using decorators. In addition, the style command is used to encapsulate shape instructions.